+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [#3915](https://github.com/bbatsov/rubocop/issues/3915): Make configurable whitelist for `Lint/SafeNavigationChain` cop. ([@pocke][])
 * [#3944](https://github.com/bbatsov/rubocop/issues/3944): Allow keyword arguments in `Style/RaiseArgs` cop. ([@mikegee][])
 * Add auto-correct to `Performance/DoubleStartEndWith`. ([@rrosenblum][])
+* [#3951](https://github.com/bbatsov/rubocop/pull/3951): Make `Rails/Date` cop to register an offence for a string without timezone. ([@sinsoku][])
 
 ### New features
 
@@ -2618,3 +2619,4 @@
 [@zverok]: https://github.com/zverok
 [@backus]: https://github.com/backus
 [@pat]: https://github.com/pat
+[@sinsoku]: https://github.com/sinsoku